requirement.txt
huggingface-hub
openai
sentence_transformers
torch
transformers
langchain
langchain-community
langchain-openai
pyyaml
diff_match_patch
spacy
nltk
wikiextractor